+++ my-app/src/App.tsx
@@ -1,9 +1,9 @@
-import { useState } from "react"
-import CardPreview from "./components/cardPreview/CardPreview";
-import CardNumberSection from "./components/cardNumberSection/CardNumberSection";
-import ExpirationDateSection from "./components/expirationDateSection/ExpirationDateSection";
-import CvcSection from "./components/cvcSection/CvcSection";
-import { AppContainer, FormLayout } from "./App.styles";
+import { useState } from 'react';
+import CardPreview from './components/CardPreview/CardPreview';
+import CardNumberSection from './components/CardNumberSection/CardNumberSection';
+import ExpirationDateSection from './components/ExpirationDateSection/ExpirationDateSection';
+import CvcSection from './components/CvcSection/CvcSection';
+import { AppContainer, FormLayout } from './App.styles';
 
 function App() {
   const [cardNumber, setCardNumber] = useState(['', '', '', '']);
@@ -14,12 +14,15 @@ function App() {
     <AppContainer>
       <CardPreview cardNumber={cardNumber} expirationDate={expirationDate} />
       <FormLayout>
-        <CardNumberSection value={cardNumber} setValue = {setCardNumber} />
-        <ExpirationDateSection value={expirationDate} setValue={setExpirationDate} />
+        <CardNumberSection value={cardNumber} setValue={setCardNumber} />
+        <ExpirationDateSection
+          value={expirationDate}
+          setValue={setExpirationDate}
+        />
         <CvcSection value={cvc} setValue={setCvc} />
       </FormLayout>
     </AppContainer>
   );
 }
 
-export default App
+export default App;
