+++ my-app/src/common/CommonSection/CommonSection.styles.ts
