+++ my-app/src/common/NumberInput/NumberInput.styles.ts
