+++ my-app/src/components/cardNumberSection/CardNumberSection.tsx
@@ -3,8 +3,8 @@ import {
   getCardNumberErrorMessage,
   isInputValidate,
 } from '../../utils/Validation';
-import CommonSection from '../common/commonSection/CommonSection';
-import NumberInput from '../common/numberInput/NumberInput';
+import CommonSection from '../../common/CommonSection/CommonSection';
+import NumberInput from '../../common/NumberInput/NumberInput';
 
 interface Props {
   value: string[];
