+++ my-app/src/components/expirationDateSection/ExpirationDateSection.tsx
@@ -4,8 +4,8 @@ import {
   getYearErrorMessage,
   isInputValidate,
 } from '../../utils/Validation';
-import CommonSection from '../common/commonSection/CommonSection';
-import NumberInput from '../common/numberInput/NumberInput';
+import CommonSection from '../../common/CommonSection/CommonSection';
+import NumberInput from '../../common/NumberInput/NumberInput';
 
 interface Props {
   value: {
